Add ability to console.dir highlighted variables
The main functionality provided by logify is the ability to highlight text and
add a `console.dir()` below the highlighted variable. I first want to add the ability
to `console.dir()` by highlighting a variable.

This PR adds the ability to highlight a section of text that will be put in `console.dir()`