Add keyboard shortcuts for `console.dir` insertion (#6)
* Add keyboard shortcuts for `console.dir`` insertion

https://eaflood.atlassian.net/browse/WATER-4884

Currently, Logify has the ability to add `console.dir()` to a user's editor. In order to compete with
other logging extensions, users typically have the option to use keyboard shortcuts to add a log. This makes
the ability to add `console.dir()` much faster and more satisfying.

This PR adds keyboard shortcuts for `console.dir()` to Logify.