Update README and prepare logify for publishing
https://eaflood.atlassian.net/browse/WATER-4884

Before placing logify onto the vscode extension marketplace, we need to
make it look all shiny :sparkles: and perfect for its big launch :rocket:. This will
mean redoing the README file properly as this is what is displayed to the users upon finding it
in vscode. We also need to add the icon for logify and any other necessary additions to make
logify look good.

This PR will update the README and prepare logify for publishing.