Add configuration settings for toggling options
https://eaflood.atlassian.net/browse/WATER-4884

As we add more and more functionality to logify, it is important we allow
users to make the experience more enjoyable by personalising logify's
functionality. Vscode allows us to do this through `configuration.settings`, meaning
we can turn on and off functionality at will.

This PR adds basic `configuration.settings` to Logify.