Add repository link to README and package.json (#11)
* Add repository link to README and package.json

https://eaflood.atlassian.net/browse/WATER-4884

Adding the repository link improves discoverability and transparency.
It allows users to easily access the source code, file issues, and contribute.
This is also recommended for publishing to the VS Code Marketplace and removes the warning from vsce publish.

This PR adds a repository link to Logify to the README and package.json