Fix incorrect name used for publisher in package.json (#14)
* Fix incorrect name used for publisher in package.json

https://eaflood.atlassian.net/browse/WATER-4884

The `publisher` field in the `package.json` was updated to match the actual publisher ID (RobleSiyad) for the extension.
This change ensures consistency between the local configuration and the VS Code Marketplace publisher account, enabling successful publishing of the extension.

This PR changes the publisher name in the package.json file for logify.