Fix console dir indentation to match surrounding code (#15)
* Fix console dir indentation to match surrounding code

https://eaflood.atlassian.net/browse/WATER-4884

After publishing Logify, it quickly became apparent that a major issue had been overlooked.
When testing, we only used variables that were at the very start of the line, and placed the
`console.dir` directly on the line underneath without considering indentation.

The problem with this is that code is often indented. This caused the inserted console statements to break
indentation conventions and look out of place, leading to a frustrating experience for users.

```js
	function greet() {
		const name = 'Alice'
	console.dir(name, { depth: null, colors: true }) // incorrect indentation
	}
```
This PR fixes the indentation of the console to match the surrounding code in Logify.